requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Desktop UI
PySide6==6.6.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Supabase
supabase==2.3.0

# Utilities
python-dotenv==1.0.0
httpx>=0.24.0,<0.25.0
aiofiles==23.2.1

# Build & Packaging
pyinstaller==6.3.0
